test: move nested redirect regression to async cluster tests
